feat: moved capacity bar to common
