+++ src/components/common/CapacityBar/index.ts
