+++ src/components/common/CapacityBar/styles.ts
@@ -9,10 +9,6 @@ export const useStyles = () =>
       flexDirection: "column",
     },
     labelText: {
-      fontFamily: montserrat.bold,
-      fontSize: RFValue(16),
-    },
-    valueText: {
       fontFamily: montserrat.medium,
       fontSize: RFValue(14),
     },
